added get_pzero.pro. This routine should compute the enigmatic pzero angle according to the observing date.
Manolo's mail to this topic states (May 17, 2018)

this pzero has always been a problem .... It should be a fixed value but it changes with time and nobody knows why. This pzero parameter is the relative orientation between the polarizer and the retarder of the main calibration unit, which is crucial (yes, crucial ) for the calibration. This year is again different ...